(#13) Added outputs merge and verification
The idea of the land-ocean ratio verification is to make sure that the NetCDF outputs have roughly as many number of land and ocean cells as we expect.
Verification script with etopo_cg=40:
  Land cells: 8509
  Ocean cells: 11971

Merging and verification scripts with outputs from etopo_cg=4:
  Land cells (is_land=1): 8622
  Ocean cells (is_land=0): 11858

Both scripts are consistent. Furthermore, the plots are sensible.